/refac move  NoteGraph dir
