+++ src/modules/NoteGraph/index.tsx
@@ -4,11 +4,12 @@ import Graph from 'react-graph-vis';
 import { useForm } from 'react-hook-form';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from 'store';
-import CreateNoteModal from '../CreateNoteModal';
+import CreateNoteModal from '../Note/CreateNoteModal';
 import '/Users/issol/mindnote/node_modules/vis-network/styles/vis-network.css';
 import { createNote, deleteNote, deleteConnection, createConnection, updateNote, updateConnection } from 'store/article/actions';
-import UpdateNoteModal from '../UpdateNoteModal';
+import UpdateNoteModal from '../Note/UpdateNoteModal';
 import UpdateConnectionModal from 'modules/Connection/UpdateConnectionModal';
+import { graphDefaultVisualOptions } from 'assets/styles/graphstyle';
 
 type EdgeDataType = {
   id: number;
@@ -19,56 +20,6 @@ type EdgeDataType = {
 
 export type NoteFormType = {
   contents: string;
-};
-
-const graphDefaultVisualOptions = {
-  autoResize: true,
-  height: '600px',
-  layout: {
-    randomSeed: 2,
-    improvedLayout: true,
-  },
-  interaction: {
-    zoomView: false,
-  },
-  edges: {
-    width: 1,
-    color: {
-      color: '#D3D3D3',
-      highlight: '#797979',
-      hover: '#797979',
-      opacity: 1.0,
-    },
-  },
-  nodes: {
-    fixed: {
-      x: false,
-      y: false,
-    },
-    widthConstraint: {
-      minimum: 1,
-      maximum: 200,
-    },
-    shape: 'box',
-    size: 10,
-    borderWidth: 1.5,
-    borderWidthSelected: 2,
-    font: {
-      color: '#adaeb9',
-      size: 15,
-      align: 'center',
-      bold: {
-        color: '#bbbdc0',
-        size: 15,
-        vadjust: 0,
-        mod: 'bold',
-      },
-    },
-    color: {
-      background: 'white',
-      border: 'black',
-    },
-  },
 };
 
 export type ConnectionFormType = {
